# Do cycle checking asynchronously