--- a/src/build/build_step.go
+++ b/src/build/build_step.go
@@ -48,7 +48,7 @@ func Build(tid int, state *core.BuildState, label core.BuildLabel, remote bool) 
 	if err := buildTarget(tid, state, target, remote); err != nil {
 		if errors.Is(err, errStop) {
 			target.SetState(core.Stopped)
-			state.LogBuildResult(tid, target.Label, core.TargetBuildStopped, "Build stopped")
+			state.LogBuildResult(tid, target, core.TargetBuildStopped, "Build stopped")
 			return
 		}
 		state.LogBuildError(tid, label, core.TargetBuildFailed, err, "Build failed: %s", err)
@@ -162,7 +162,7 @@ func buildTarget(tid int, state *core.BuildState, target *core.BuildTarget, runR
 		log.Debug("Finished pre-build function for %s", target.Label)
 	}
 
-	state.LogBuildResult(tid, target.Label, core.TargetBuilding, "Preparing...")
+	state.LogBuildResult(tid, target, core.TargetBuilding, "Preparing...")
 	if state.PrepareOnly && state.IsOriginalTarget(target) {
 		return prepareOnly(tid, state, target)
 	}
@@ -219,7 +219,7 @@ func buildTarget(tid int, state *core.BuildState, target *core.BuildTarget, runR
 					copyFilegroupHashes(state, target)
 				}
 				target.SetState(core.Reused)
-				state.LogBuildResult(tid, target.Label, core.TargetCached, "Unchanged")
+				state.LogBuildResult(tid, target, core.TargetCached, "Unchanged")
 				buildLinks(state, target)
 				return nil // Nothing needs to be done.
 			}
@@ -234,10 +234,10 @@ func buildTarget(tid int, state *core.BuildState, target *core.BuildTarget, runR
 				return err
 			} else if changed {
 				target.SetState(core.Built)
-				state.LogBuildResult(tid, target.Label, core.TargetBuilt, "Built")
+				state.LogBuildResult(tid, target, core.TargetBuilt, "Built")
 			} else {
 				target.SetState(core.Unchanged)
-				state.LogBuildResult(tid, target.Label, core.TargetCached, "Unchanged")
+				state.LogBuildResult(tid, target, core.TargetCached, "Unchanged")
 			}
 			buildLinks(state, target)
 			return nil
@@ -278,12 +278,12 @@ func buildTarget(tid int, state *core.BuildState, target *core.BuildTarget, runR
 		if err := target.CheckSecrets(); err != nil {
 			return err
 		}
-		state.LogBuildResult(tid, target.Label, core.TargetBuilding, "Preparing...")
+		state.LogBuildResult(tid, target, core.TargetBuilding, "Preparing...")
 		if err := prepareSources(state.Graph, target); err != nil {
 			return fmt.Errorf("Error preparing sources for %s: %s", target.Label, err)
 		}
 
-		state.LogBuildResult(tid, target.Label, core.TargetBuilding, target.BuildingDescription)
+		state.LogBuildResult(tid, target, core.TargetBuilding, target.BuildingDescription)
 		metadata, err = buildMaybeRemotely(state, target, cacheKey)
 		if err != nil {
 			return err
@@ -323,10 +323,10 @@ func buildTarget(tid int, state *core.BuildState, target *core.BuildTarget, runR
 	if runRemotely {
 		if metadata.Cached {
 			target.SetState(core.ReusedRemotely)
-			state.LogBuildResult(tid, target.Label, core.TargetBuilt, "Reused existing action")
+			state.LogBuildResult(tid, target, core.TargetBuilt, "Reused existing action")
 		} else {
 			target.SetState(core.BuiltRemotely)
-			state.LogBuildResult(tid, target.Label, core.TargetBuilt, "Built remotely")
+			state.LogBuildResult(tid, target, core.TargetBuilt, "Built remotely")
 		}
 		if state.ShouldDownload(target) {
 			if err := state.EnsureDownloaded(target); err != nil {
@@ -339,7 +339,7 @@ func buildTarget(tid int, state *core.BuildState, target *core.BuildTarget, runR
 		return fmt.Errorf("failed to store target build metadata for %s: %w", target.Label, err)
 	}
 
-	state.LogBuildResult(tid, target.Label, core.TargetBuilding, "Collecting outputs...")
+	state.LogBuildResult(tid, target, core.TargetBuilding, "Collecting outputs...")
 	outs, outputsChanged, err := moveOutputs(state, target)
 	if err != nil {
 		return fmt.Errorf("error moving outputs for target %s: %w", target.Label, err)
@@ -354,7 +354,7 @@ func buildTarget(tid int, state *core.BuildState, target *core.BuildTarget, runR
 	}
 	buildLinks(state, target)
 	if state.Cache != nil {
-		state.LogBuildResult(tid, target.Label, core.TargetBuilding, "Storing...")
+		state.LogBuildResult(tid, target, core.TargetBuilding, "Storing...")
 		newCacheKey := mustShortTargetHash(state, target)
 
 		// If the build could modify the target, store the metadata in the cache based on the original state of the
@@ -376,9 +376,9 @@ func buildTarget(tid int, state *core.BuildState, target *core.BuildTarget, runR
 		}
 	}
 	if outputsChanged {
-		state.LogBuildResult(tid, target.Label, core.TargetBuilt, "Built")
+		state.LogBuildResult(tid, target, core.TargetBuilt, "Built")
 	} else {
-		state.LogBuildResult(tid, target.Label, core.TargetBuilt, "Built (unchanged)")
+		state.LogBuildResult(tid, target, core.TargetBuilt, "Built (unchanged)")
 	}
 	return nil
 }
@@ -427,10 +427,10 @@ func retrieveArtifacts(tid int, state *core.BuildState, target *core.BuildTarget
 	// later tries to add more outputs.
 	if len(target.DeclaredOutputs()) == 0 && len(target.DeclaredNamedOutputs()) == 0 {
 		target.SetState(core.Unchanged)
-		state.LogBuildResult(tid, target.Label, core.TargetCached, "Nothing to do")
+		state.LogBuildResult(tid, target, core.TargetCached, "Nothing to do")
 		return true
 	}
-	state.LogBuildResult(tid, target.Label, core.TargetBuilding, "Checking cache...")
+	state.LogBuildResult(tid, target, core.TargetBuilding, "Checking cache...")
 
 	cacheKey := mustShortTargetHash(state, target)
 
@@ -449,10 +449,10 @@ func retrieveArtifacts(tid int, state *core.BuildState, target *core.BuildTarget
 			return false
 		} else if oldOutputHash == nil || !bytes.Equal(oldOutputHash, newOutputHash) {
 			target.SetState(core.Cached)
-			state.LogBuildResult(tid, target.Label, core.TargetCached, "Cached")
+			state.LogBuildResult(tid, target, core.TargetCached, "Cached")
 		} else {
 			target.SetState(core.Unchanged)
-			state.LogBuildResult(tid, target.Label, core.TargetCached, "Cached (unchanged)")
+			state.LogBuildResult(tid, target, core.TargetCached, "Cached (unchanged)")
 		}
 		buildLinks(state, target)