--- a/src/core/BUILD
+++ b/src/core/BUILD
@@ -179,12 +179,6 @@ go_test(
 )
 
 benchmark(
-    name = "cycle_detector_benchmark",
-    srcs = ["cycle_detector_benchmark_test.go"],
-    deps = [":core"],
-)
-
-benchmark(
     name = "build_target_benchmark",
     srcs = ["build_target_benchmark_test.go"],
     deps = [":core"],