--- a/src/core/cycle_detector.go
+++ b/src/core/cycle_detector.go
@@ -6,40 +6,40 @@ import (
 )
 
 type cycleDetector struct {
-	state *BuildState
+	graph *BuildGraph
 }
 
 // Check runs a single check of the build graph to see if any cycles can be detected.
-// If it finds one it logs an async error to the state object.
-func (c *cycleDetector) Check() error {
+// If it finds one an errCycle is returned.
+func (c *cycleDetector) Check() *errCycle {
 	permanent := map[*BuildTarget]struct{}{}
 	temporary := map[*BuildTarget]struct{}{}
 
-	var visit func(target *BuildTarget) []*BuildTarget
-	visit = func(target *BuildTarget) []*BuildTarget {
+	var visit func(target *BuildTarget) ([]*BuildTarget, bool)
+	visit = func(target *BuildTarget) ([]*BuildTarget, bool) {
 		if _, present := permanent[target]; present {
-			return nil
+			return nil, false
 		} else if _, present := temporary[target]; present {
-			return []*BuildTarget{target}
+			return []*BuildTarget{target}, false
 		}
 		temporary[target] = struct{}{}
 		for _, dep := range target.Dependencies() {
-			if cycle := visit(dep); cycle != nil {
-				return append([]*BuildTarget{target}, cycle...)
+			if cycle, done := visit(dep); cycle != nil {
+				if done || target == cycle[len(cycle)-1] {
+					return cycle, true  // This target is already in the cycle
+				}
+				return append([]*BuildTarget{target}, cycle...), false
 			}
 		}
 		delete(temporary, target)
 		permanent[target] = struct{}{}
-		return nil
+		return nil, false
 	}
 
-	for _, target := range c.state.Graph.AllTargets() {
+	for _, target := range c.graph.AllTargets() {
 		if _, present := permanent[target]; !present {
-			if cycle := visit(target); cycle != nil {
-				err := errCycle{Cycle: cycle}
-				c.state.LogBuildError(0, cycle[0].Label, TargetBuildFailed, err, "")
-				c.state.Stop()
-				return err
+			if cycle, _ := visit(target); cycle != nil {
+				return &errCycle{Cycle: cycle}
 			}
 		}
 	}
@@ -51,10 +51,11 @@ type errCycle struct{
 	Cycle []*BuildTarget
 }
 
-func (err errCycle) Error() string {
+func (err *errCycle) Error() string {
 	labels := make([]string, len(err.Cycle))
 	for i, t := range err.Cycle {
 		labels[i] = t.Label.String()
 	}
+	labels = append(labels, labels[0])
 	return fmt.Sprintf("Dependency cycle found:\n%s\nSorry, but you'll have to refactor your build files to avoid this cycle", strings.Join(labels, "\n -> "))
 }