--- a/src/core/cycle_detector.go
+++ b/src/core/cycle_detector.go
@@ -13,17 +13,22 @@ type cycleDetector struct {
 // If it finds one an errCycle is returned.
 func (c *cycleDetector) Check() *errCycle {
 	log.Debug("Running cycle detection...")
-	permanent := map[*BuildTarget]struct{}{}
-	temporary := map[*BuildTarget]struct{}{}
+	complete := map[*BuildTarget]struct{}{}
+	partial := map[*BuildTarget]struct{}{}
 
+	// visit visits a target and all its transitive dependencies. As each is visited they are marked as
+	// partially visited; when we bottom out a tree successfully we mark it as completely visited (this
+	// saves us from revisiting any node we've successfully visited before).
+	// If a cycle is found it returns a slice of the targets in that cycle, and a bool indicating if the
+	// cycle is complete or not (if not the caller will need to add its node to it as well).
 	var visit func(target *BuildTarget) ([]*BuildTarget, bool)
 	visit = func(target *BuildTarget) ([]*BuildTarget, bool) {
-		if _, present := permanent[target]; present {
+		if _, present := complete[target]; present {
 			return nil, false
-		} else if _, present := temporary[target]; present {
+		} else if _, present := partial[target]; present {
 			return []*BuildTarget{target}, false
 		}
-		temporary[target] = struct{}{}
+		partial[target] = struct{}{}
 		for _, dep := range target.Dependencies() {
 			if cycle, done := visit(dep); cycle != nil {
 				if done || target == cycle[len(cycle)-1] {
@@ -32,13 +37,13 @@ func (c *cycleDetector) Check() *errCycle {
 				return append([]*BuildTarget{target}, cycle...), false
 			}
 		}
-		delete(temporary, target)
-		permanent[target] = struct{}{}
+		delete(partial, target)
+		complete[target] = struct{}{}
 		return nil, false
 	}
 
 	for _, target := range c.graph.AllTargets() {
-		if _, present := permanent[target]; !present {
+		if _, present := complete[target]; !present {
 			if cycle, _ := visit(target); cycle != nil {
 				log.Debug("Cycle detection complete, cycle found: %s", cycle)
 				return &errCycle{Cycle: cycle}