This moves cycle checking from happening always to only at a point where we have no active work happening for a few seconds. That means there's a little pause before it'll find the cycle but I don't think that's a big problem.

Most of the guts of it is in `forwardResults()` which intercepts all the parse/build/test updates to add this timing logic. The cycle detector can now use a very simple synchronous algorithm since it should never happen on the hot path.

On my machine this reduces parse times from ~10s to ~6.5s (specifically 9.56s, 9.88s, 12.33s -> 6.50s, 6.54s, 6.61) vs. current head of master.

The in-repo cyclic dependency test passes; can give it a go if we have any other tests around this stuff too.